refactor, added chat component and change code style
