--- src/bus/client/clientMessage/reducers.ts
@@ -1,10 +1,0 @@
-// Types
-import * as types from './types';
-
-export const setClientMessage: types.BaseContact<types.ClientMessage> = (__, action) => {
-    return action.payload;
-};
-
-export const closeClientMessage = () => {
-    return null;
-};
