+++ src/bus/client/selectedMessage/reducers.ts
@@ -1,0 +1,10 @@
+// Types
+import * as types from './types';
+
+export const setSelectedMessage: types.BaseContact<types.SelectedMessage> = (__, action) => {
+    return action.payload;
+};
+
+export const closeSelectedMessage = () => {
+    return null;
+};
