+++ src/bus/client/selectedMessage/slice.ts
@@ -1,0 +1,20 @@
+// Core
+import { createSlice } from '@reduxjs/toolkit';
+
+// Types
+import * as types from './types';
+
+// Reducers
+import * as reducers from './reducers';
+
+const initialState = null;
+
+export const SelectedMessageSlice = createSlice<types.SelectedMessageState, typeof reducers>({
+    name: 'SelectedMessage',
+    initialState,
+    reducers,
+});
+
+export const sliceName = SelectedMessageSlice.name;
+export const SelectedMessageActions = SelectedMessageSlice.actions;
+export default SelectedMessageSlice.reducer;
