+++ src/bus/client/selectedMessage/types.ts
@@ -1,0 +1,12 @@
+// Core
+import { CaseReducer, PayloadAction } from '@reduxjs/toolkit';
+
+// Types
+import { Message } from '../../messages/types';
+
+// State
+export type SelectedMessage = Message
+export type SelectedMessageState = SelectedMessage | null
+
+// Contracts
+export type BaseContact<T = any> = CaseReducer<SelectedMessageState, PayloadAction<T>>
