+++ src/init/redux/index.ts
@@ -6,7 +6,7 @@ import togglers from '../../bus/client/togglers';
 import user from '../../bus/user/slice';
 import messages from '../../bus/messages/slice';
 import message from '../../bus/messages/slice';
-import clientMessage from '../../bus/client/clientMessage/slice';
+import selectedMessage from '../../bus/client/selectedMessage/slice';
 import keyboard from '../../bus/client/keyboard/slice';
 // import __entityName__ from '../../bus/__entityName__/slice';
 
@@ -22,7 +22,7 @@ export const store = configureStore({
         user,
         messages,
         message,
-        clientMessage,
+        selectedMessage,
         keyboard,
         // __entityName__,
     },
