+++ src/view/components/RegistrationForm/index.tsx
@@ -1,18 +1,33 @@
 // Core
-import React, { FC } from 'react';
+import React, { FC, useState } from 'react';
+
+// Bus
+import { useUser } from '../../../bus/user';
 
 // Styles
 import * as S from './styles';
 
-// Types
-import { FormPropTypes } from '../ChatForm';
+export const RegistrationForm: FC = () => {
+    const { registerUser } = useUser();
 
-export const RegistrationForm: FC<FormPropTypes> = ({ onSubmitForm, inputValue, onChangeInput }) => {
+    const randomNumbers = () => Math.floor(1000 + (Math.random() * 9000));
+    const [ username, setUsername ] = useState(`NINJA:${randomNumbers()}`);
+
+    const handleChangeInput = (event: React.ChangeEvent<HTMLInputElement>) => {
+        return setUsername(event.target.value);
+    };
+
+    const handleSubmitForm = (event: React.FormEvent<HTMLFormElement>, username: string) => {
+        event.preventDefault();
+
+        registerUser(username);
+    };
+
     return (
         <S.Container>
             <S.Form
                 action = '#'
-                onSubmit = { onSubmitForm }>
+                onSubmit = { (event) => void handleSubmitForm(event, username) }>
                 <S.Label htmlFor = 'text'>
                     <S.LabelText>
                         Enter your NinjaName:
@@ -21,12 +36,12 @@ export const RegistrationForm: FC<FormPropTypes> = ({ onSubmitForm, inputValue, 
                         name = 'text'
                         placeholder = 'Write your ninja name'
                         type = 'text'
-                        value = { inputValue }
-                        onChange = { onChangeInput }
+                        value = { username }
+                        onChange = { handleChangeInput }
                     />
                 </S.Label>
                 <S.SubmitBtn
-                    disabled = { !inputValue }
+                    disabled = { !username }
                     type = 'submit'>
                     Submit
                 </S.SubmitBtn>
