+++ src/view/components/index.ts
@@ -1,6 +1,10 @@
 export * from './ErrorBoundary';
+export * from './OwnerInfo';
+export * from './CreateMessageForm';
 export * from './UserInfo';
-export * from './ChatForm';
-export * from './RegistrationUserInfo';
 export * from './RegistrationForm';
+export * from './EditMessageForm';
+export * from './Chat';
+export * from './MessageInfoActions';
+export * from './MessageDateInfo';
 export * from './Keyboard';
