+++ src/view/pages/Registration/index.tsx
@@ -1,8 +1,5 @@
 // Core
-import React, { FC, useState } from 'react';
-
-// Bus
-import { useUser } from '../../../bus/user';
+import React, { FC } from 'react';
 
 // Components
 import { ErrorBoundary, RegistrationUserInfo, RegistrationForm } from '../../components';
@@ -10,33 +7,11 @@ import { ErrorBoundary, RegistrationUserInfo, RegistrationForm } from '../../com
 // Styles
 import * as S from './styles';
 
-// Types
-import { FormChangeEvent, FormSubmitEvent } from '../../components';
-
 const Registration: FC = () => {
-    const { registerUser } = useUser();
-
-    const random = () => Math.floor(1000 + (Math.random() * 9000));
-    const [ username, setUsername ] = useState(`NINJA:${random()}`);
-
-    const handleChangeInput = (event: FormChangeEvent) => {
-        return setUsername(event.target.value);
-    };
-
-    const handleSubmitForm = (event: FormSubmitEvent, username: string) => {
-        event.preventDefault();
-
-        registerUser(username);
-    };
-
     return (
         <S.Container>
             <RegistrationUserInfo />
-            <RegistrationForm
-                inputValue = { username }
-                onChangeInput = { handleChangeInput }
-                onSubmitForm = { (event) => handleSubmitForm(event, username) }
-            />
+            <RegistrationForm/>
         </S.Container>
     );
 };
